pyproject.toml
[tool.poetry]
name = "pytorch_fastapi_project"
version = "0.1.0"
description = "Consolidated PyTorch and FastAPI project for AWS deployment and GHA testing"
authors = ["soutrik71 <soutrik.chowdhury@ab-inbev.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.6"
black = "24.8.0"
coverage = ">=7.6.1"
hydra-colorlog = "1.2.0"
hydra-core = "1.3.2"
lightning = {version = "2.4.0", extras = ["extra"]}
loguru = "0.7.2"
pytest = "^8.3.3"
rich = "13.8.1"
rootutils = "1.0.7"
tensorboard = "2.17.1"
timm = "1.0.9"
pandas = "^2.2.3"
numpy = "^1.26.0"
torch = {version = "^2.4.1", source = "pytorch_cuda"}
torchvision = {version = "^0.19.1", source = "pytorch_cuda"}
torchaudio = {version = "^2.4.1", source = "pytorch_cuda"}
seaborn = "^0.13.2"
pydantic = "^2.9.2"
kaggle = "^1.6.17"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
flake8 = "^7.1.1"
dvc-gdrive = "^3.0.1"
dvc-azure = "^3.1.0"
transformers = "^4.45.2"
fastapi = "^0.115.4"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.0"
tenacity = "^9.0.0"
gunicorn = "^23.0.0"
aim = "^3.25.0"
mlflow = "^2.17.1"
hydra-optuna-sweeper = "^1.2.0"
dvc = "^3.56.0"
platformdirs = "3.10"
fastapi-utils = "^0.7.0"
httpx = "^0.27.2"
typing-inspect = "^0.9.0"
requests = "^2.32.3"
fastapi-restful = {extras = ["all"], version = "^0.6.0"}
aioredis = "^2.0.1"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
confluent-kafka = "^2.6.0"
aiokafka = "^0.12.0"
azure-servicebus = "^7.12.3"
aiohttp = "^3.10.10"
aiofiles = "*"
aiologger = "^0.7.0"
pyyaml = "^6.0.2"
sqlalchemy-utils = "^0.41.2"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
fastapi-limiter = "^0.1.6"
redis = "5.0.8"
redisearch = "2.0.0"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
fastapi-cache2 = "^0.2.2"
aiocache = "^0.12.3"
dvc-s3 = "^3.2.0"
litserve = "^0.2.4"
gpustat = "^1.1.1"
nvitop = "^1.3.2"
gradio = "5.7.1"
gradio-client = "^1.5.0"
accelerate = "^1.1.1"
pyopenssl = "<23.0.0"
cryptography = "^44.0.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.20.3"

[[tool.poetry.source]]
name = "pytorch_cuda"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"